Refactored asset size type to Long

+++ logic/src/commonMain/kotlin/com/wire/kalium/logic/data/message/AssetContent.kt
@@ -1,7 +1,7 @@
 package com.wire.kalium.logic.data.message
 
 data class AssetContent(
-    val sizeInBytes: Int,
+    val sizeInBytes: Long,
     val name: String? = null,
     val mimeType: String,
     val metadata: AssetMetadata? = null,
